[explorer/rest] feat: Added unit test helper function
