+++ explorer/rest/tests/test/DatabaseTestUtils.py
@@ -19,7 +19,7 @@ from rest.model.Statistic import (
 	StatisticTransactionDateRangeView,
 	StatisticTransactionView
 )
-from rest.model.Transaction import TransactionView
+from rest.model.Transaction import TransactionQuery, TransactionView
 
 Block = namedtuple(
 	'Block',
@@ -1132,3 +1132,16 @@ class DatabaseTestBase(unittest.TestCase):
 
 	def tearDown(self):
 		self.postgresql.stop()
+
+	@staticmethod
+	def _make_transaction_query(**kwargs):
+		defaults = TransactionQuery(
+			height=None,
+			transaction_types=None,
+			sender=None,
+			address=None,
+			sender_address=None,
+			recipient_address=None,
+			mosaic=None
+		)
+		return defaults._replace(**kwargs)
